third_party/WebKit/LayoutTests/fast/dom/Element/offset-parent-will-change.html

<!DOCTYPE html>
<div id="will-change-transform" style="will-change:transform;">
    <div id="test"></div>
</div>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
    test(function() {
       assert_equals(document.getElementById("test").offsetParent, document.getElementById("will-change-transform"));
    }, "inside will-change:transform");
</script>
